test(web): extract shared app suite fixtures
